requirements.txt
gradio
transformers
torch
sentence-transformers
datasets
duckduckgo-search
tldextract
requests
python-dotenv
azure-ai-inference
azure-core
openai
pdfminer.six
playwright
playwright-stealth
aiohttp
underthesea
tiktoken
sentencepiece